Uploaded program's .py file
Initial commit/upload for a project I am working on and will update later. Commit includes an uploaded python file that holds the code of the program itself. No anagram code yet, just the first of likely several algorithms to be used on several sites to retrieve names of figures in mythology.