AAAP v2.0
Significant changes to source code to give AAAP a tkinter gui and better overall organization. Added a folder of wireframe CAD drawings for each screen in AAAP that were used to aid in the creation of each screen frame layout of the program. Also added a folder for assets used in the project.